+++ src/components/Accordion/Accordion.stories.tsx
@@ -1,12 +1,14 @@
 import * as React from "react";
 import { Meta, Story } from "@storybook/react/types-6-0";
+import range from "lodash/range";
 import Accordion, { AccordionProps } from "./Accordion";
 import { action } from "@storybook/addon-actions";
 import { withDesign } from "storybook-addon-designs";
 
-import { BasicCheckbox } from "../Checkbox/Checkbox.stories";
-import List from "../List/List";
-import { ListTypes } from "../List/ListTypes";
+// import List from "../List/List";
+// import { ListTypes } from "../List/ListTypes";
+import { list as ListStory } from "../List/List.stories";
+import Link from "../Link/Link";
 
 export default {
     title: "Accordion",
@@ -16,20 +18,12 @@ export default {
 
 // Set up the reusable template to create multiple stories for the
 // Accordian component.
-const AccordionTemplate = ({ items, ...args }) => (
-    <Accordion {...args}>
-        <List
-            type={ListTypes.Unordered}
-            id="checkbox-list"
-            modifiers={["no-list-styling"]}
-        >
-            {items.map((item) => (
-                <li>
-                    <BasicCheckbox {...item} />
-                </li>
-            ))}
-        </List>
-    </Accordion>
+const AccordionTemplate = ({ count, children, ...args }) => (
+    <>
+        {range(count).map((i) => (
+            <Accordion {...args}>{children}</Accordion>
+        ))}
+    </>
 );
 
 // This is one specific story where it will render a list of checkboxes.
@@ -40,7 +34,8 @@ export const AccordionWithCheckboxes = AccordionTemplate.bind({});
 AccordionWithCheckboxes.args = {
     accordionLabel: "Click to expand",
     labelId: "accordionBtn",
-    items: [BasicCheckbox.args],
+    children: ListStory,
+    count: 1,
 };
 
 AccordionWithCheckboxes.argTypes = {
@@ -48,6 +43,7 @@ AccordionWithCheckboxes.argTypes = {
     className: { table: { disable: true } },
     id: { table: { disable: true } },
     modifiers: { table: { disable: true } },
+    children: { table: { disable: true } },
 };
 
 AccordionWithCheckboxes.parameters = {
@@ -58,16 +54,41 @@ AccordionWithCheckboxes.parameters = {
     },
 };
 
+const faqContent = (
+    <>
+        <p>
+            Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do
+            eiusmod tempor incididunt ut labore et dolore magna aliqua. Lorem
+            ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod
+            tempor incididunt ut labore et dolore magna aliqua. Lorem ipsum
+            dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor
+            incididunt ut labore et dolore magna aliqua.
+        </p>
+
+        <p>
+            <Link href="#">Learn More</Link>
+        </p>
+    </>
+);
+
 // This is one specific story where it will render a list of FAQs.
-export const AccordionAsFAQ = AccordionTemplate.bind({});
-AccordionAsFAQ.args = {
+export const AccordionAsFAQSet = AccordionTemplate.bind({});
+AccordionAsFAQSet.args = {
     accordionLabel: "FAQ Question Lorem Ipsum",
     modifiers: ["faq"],
     labelId: "accordionBtn",
-    items: [BasicCheckbox.args],
+    children: faqContent,
+    count: 3,
 };
 
-AccordionAsFAQ.parameters = {
+AccordionAsFAQSet.argTypes = {
+    blockName: { table: { disable: true } },
+    className: { table: { disable: true } },
+    id: { table: { disable: true } },
+    children: { table: { disable: true } },
+};
+
+AccordionAsFAQSet.parameters = {
     design: {
         type: "figma",
         url: